requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
pytest
PyYAML
